% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodpanda_feature_change.R
\name{analyze_feature_changes}
\alias{analyze_feature_changes}
\title{Compare two Foodpanda data frame for preliminary feature change analysis}
\usage{
analyze_feature_changes(df_before, df_after, removeFeatures = "")
}
\arguments{
\item{df_before}{a data frame from foodpanda}

\item{df_after}{a data frame fro foodpanda}

\item{removeFeatures}{a character of feature names that don't need to be analyzed (default = '')}
}
\value{
a list of analyzed result
}
\description{
Compare two Foodpanda data frame for preliminary feature change analysis
}
