% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodpanda_feature_change.R
\name{fix_jsonChar}
\alias{fix_jsonChar}
\title{Fix menu json string error in old data}
\usage{
fix_jsonChar(mm)
}
\arguments{
\item{mm}{a string of menu json}
}
\value{
a string of correct json
}
\description{
Fix menu json string error in old data
}
