% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodpanda.R
\name{FoodPanda}
\alias{FoodPanda}
\title{Initiate a Foodpanda Data instance for preliminary foodpada data import and analysis}
\usage{
FoodPanda(
  shopListFolderUrl =
    "https://drive.google.com/drive/folders/1gbUGUggYxvUBzyMyW1et4UZLQ4UguUCk",
  shopMenuFolderUrl =
    "https://drive.google.com/drive/folders/14Yk8kBRcWrULfI5vqI3508eyJNMH9RtM"
)
}
\arguments{
\item{shopListFolderUrl}{a character vector}

\item{shopMenuFolderUrl}{a character vector}
}
\value{
an R6 Data Foodpanda class object
}
\description{
Initiate a Foodpanda Data instance for preliminary foodpada data import and analysis
}
