% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{get_popular_items_from_menuJson}
\alias{get_popular_items_from_menuJson}
\title{Get popular items from menu json}
\usage{
get_popular_items_from_menuJson(menu)
}
\arguments{
\item{menu}{a list from jsonlite::fromJSON()}
}
\value{
a character vector of popular items
}
\description{
Get popular items from menu json
}
