% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Get weather data from all Taiwan weather observation stations}
\usage{
get_weather_data(date)
}
\arguments{
\item{date}{a string of "YYYY-MM-DD" format for day weather, "YYYY-MM" for month weather, "YYYY" for year weather.}
}
\value{
a list of weather data from all observation stations. Check weather_stations data frame for more details.
}
\description{
Get weather data from all Taiwan weather observation stations
}
