% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodpanda_feature_change.R
\name{kable_clip}
\alias{kable_clip}
\title{A wrapper of knitr::kable}
\usage{
kable_clip(...)
}
\arguments{
\item{...}{parameters passed to kable}
}
\value{
pass kable return to clipboard for pasting somewhere else
}
\description{
A wrapper of knitr::kable
}
