% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodpanda_feature_change.R
\name{parse_menu_to_list_of_dataframe}
\alias{parse_menu_to_list_of_dataframe}
\title{Parse menu feature into list of menu data frames}
\usage{
parse_menu_to_list_of_dataframe(df)
}
\arguments{
\item{df}{A data frame with menu column as string of Json}
}
\value{
a list of menu data frames
}
\description{
Parse menu feature into list of menu data frames
}
